webhook/serializers/destination.py
from rest_framework import serializers
from webhook.models import Destination

class DestinationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Destination
        fields = '__all__'
