requirements.txt
# Django
Django==5.0.1
# Django Helpers
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
django-choices==2.0.0
# Database
psycopg2-binary==2.9.9
# Request
requests==2.31.0
